+++ processor/src/test/resources/bad_input/com/example/package_access/Observable2Model.java
@@ -1,0 +1,10 @@
+package com.example.package_access;
+
+import arez.annotations.ArezComponent;
+import com.example.package_access.other.BaseObservable2Model;
+
+@ArezComponent
+abstract class Observable2Model
+  extends BaseObservable2Model
+{
+}