+++ processor/src/test/resources/bad_input/com/example/package_access/Observable3Model.java
@@ -1,0 +1,10 @@
+package com.example.package_access;
+
+import arez.annotations.ArezComponent;
+import com.example.package_access.other.BaseObservable3Model;
+
+@ArezComponent
+abstract class Observable3Model
+  extends BaseObservable3Model
+{
+}