+++ processor/src/test/resources/bad_input/com/example/package_access/other/BaseObservable2Model.java
@@ -1,0 +1,11 @@
+package com.example.package_access.other;
+
+import arez.annotations.Observable;
+
+public abstract class BaseObservable2Model
+{
+  @Observable
+  public abstract String getMyValue();
+
+  abstract void setMyValue( String value );
+}