+++ processor/src/test/resources/bad_input/com/example/package_access/other/BaseObservable3Model.java
@@ -1,0 +1,11 @@
+package com.example.package_access.other;
+
+import arez.annotations.Observable;
+
+public abstract class BaseObservable3Model
+{
+  abstract String getMyValue();
+
+  @Observable
+  public abstract void setMyValue( String value );
+}